+++ src/Body.js
@@ -13,9 +13,9 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-import React, { useContext, useCallback } from 'react'
+import React, { useContext, useCallback, useRef, useEffect } from 'react'
 import Row from './Row'
-import { TableDispatch, SELECTING } from './actions'
+import { TableDispatch, SELECTING, VSCROLL } from './actions'
 import PropTypes from 'prop-types'
 import {
   TableStateType,
@@ -26,9 +26,9 @@ import {
 import './Body.css'
 
 const Body = props => {
-  // console.log('Body', props)
+  console.log('Body', props)
   const { state, components, layouts, colOrder, labels, rowIdAttr } = props
-  const { data, columns, selectedIds } = state
+  const { data, columns, selectedIds, scrollTop = 0 } = state
   const dispatch = useContext(TableDispatch)
   const handleCellCheckChange = useCallback(
     event => {
@@ -47,8 +47,23 @@ const Body = props => {
     },
     [selectedIds, dispatch]
   )
+  const handleScroll = useCallback(event => {
+    dispatch({ type: VSCROLL, scrollTop: event.target.scrollTop })
+  }, [])
+  const ref = useRef(null)
+  useEffect(() => {
+    const { current } = ref
+    if (scrollTop > 0 && current) {
+      current.scrollTop = scrollTop
+    }
+  })
   return (
-    <div className='rrt-tbody' onChange={handleCellCheckChange}>
+    <div
+      className='rrt-tbody'
+      ref={ref}
+      onChange={handleCellCheckChange}
+      onScroll={handleScroll}
+    >
       {data.map((row, index) => {
         // console.log('row', row)
         const id = row[rowIdAttr]
@@ -87,7 +102,8 @@ export const areEqual = (prev, next) => {
   const areEqual =
     prev.colOrder === next.colOrder &&
     prevState.data === nextState.data &&
-    prevState.selectedIds === nextState.selectedIds
+    prevState.selectedIds === nextState.selectedIds &&
+    prevState.scrollTop === nextState.scrollTop
   /* if (!areEqual) {
     console.log('!Body.areEqual')
   } */