+++ src/Table.js
@@ -63,6 +63,7 @@ const resizerReducer = (state, action) => {
  * | COLUMN_RESIZING   | `<ColumnsType>` | Triggered when the user resizes a column   |
  * | SELECTING   | `<SelectionType>` | Triggered when the user changes the row selection |
  * | SORTING   | `<SortType>` | Triggered when the user changes table sorting |
+ * | VSCROLL | `<VScrollType>` | Triggered when the table body is scrolled vertically |
  *
  * `<PagingType>`
  *
@@ -89,6 +90,11 @@ const resizerReducer = (state, action) => {
  * |------------|--------------|--------------------------|
  * | sort  | `<string>` | The column id used for sorting, prefixed by '+' for ascending sort or '-' for descending sort (used only for sorting)|
  *
+ * `<VScrollType>`
+ *
+ * | Key        | Type         | Description                                                                        |
+ * |------------|--------------|--------------------------|
+ * | scrollTop  | `<number>` | the offset to the top of the table body element |
  */
 const Table = props => {
   // console.log('Table', props)
@@ -203,6 +209,7 @@ Table.propTypes = {
    * | sort | `<string>` | The column id used for sorting, prefixed by '+' for ascending sort or '-' for descending sort (used only for sorting)
    * | selectedIds | `<object>` | a Set of ids currently selected (used only for selection)
    * | loading | `<bool>` | True if the table is loading its data
+   * | scrollTop | `<number>` | The scrolling offset to apply initially to the table body
    *
    * `<ColumnType>` is an object, which contains the following keys:
    *
@@ -210,12 +217,12 @@ Table.propTypes = {
    * |-----------------|-------------------|-----------------------------------------------------|
    * | id              | `<string>`        | Unique id identifying the column
    * | label           | `<string>`        | The label to display for this column
-   * | resizable       | `<bool>           | True if the column can be resized
-   * | sortable        | `<bool>           | True if the column can be sorted
-   * | minWidth        | `<number>         | The min width of the column (in pixels)
-   * | width           | `<number>         | The default width of the column (in pixels)
-   * | Cell            | `<elementType>    | The React component to use for cells corresponding to this column
-   * | Filter          | `<elementType>    | A React component to use to specify a filter is the column can be filtered
+   * | resizable       | `<bool>`          | True if the column can be resized
+   * | sortable        | `<bool>`          | True if the column can be sorted
+   * | minWidth        | `<number>`        | The min width of the column (in pixels)
+   * | width           | `<number>`        | The default width of the column (in pixels)
+   * | Cell            | `<elementType>`   | The React component to use for cells corresponding to this column
+   * | Filter          | `<elementType>`   | A React component to use to specify a filter is the column can be filtered
    */
   state: TableStateType,
   /**