+++ src/prop-types.js
@@ -44,7 +44,8 @@ export const TableStateType = PropTypes.shape({
   total: PropTypes.number,
   loading: PropTypes.bool,
   sort: PropTypes.string,
-  selectedIds: PropTypes.object
+  selectedIds: PropTypes.object,
+  scrollTop: PropTypes.number
 })
 
 export const LabelsType = PropTypes.shape({
@@ -81,5 +82,6 @@ export const IconType = PropTypes.oneOf([
   'previous',
   'next',
   'last',
-  'cancel'
+  'cancel',
+  'sortable'
 ])