+++ stories/people/peopleReducer.js
@@ -31,7 +31,7 @@ export const DELETING = 'DELETING'
 export const QUERYING = 'QUERYING'
 
 export const tableReducer = (state, action) => {
-  console.log('tableReducer', { state, action })
+  console.log('tableReducer', state, action)
   const { columns } = state
   const {
     type,
@@ -46,7 +46,7 @@ export const tableReducer = (state, action) => {
     total,
     dispatch
   } = action
-  switch (action.type) {
+  switch (type) {
     case LOADING: {
       const fetchIdRef = state.fetchIdRef
       const fetchId = ++fetchIdRef.current
@@ -56,7 +56,7 @@ export const tableReducer = (state, action) => {
     case END_LOADING: {
       return { ...state, loading: false, data, pageCount, total }
     }
-    case DELETING:
+    case DELETING: {
       const fetchIdRef = state.fetchIdRef
       const fetchId = ++fetchIdRef.current
       const newState = {
@@ -67,6 +67,7 @@ export const tableReducer = (state, action) => {
       }
       getModels({ fetchId, dispatch, state: newState })
       return newState
+    }
     case SORTING:
       return { ...state, sort }
     case FILTERING:
@@ -94,7 +95,7 @@ export const tableReducer = (state, action) => {
       return action.state
     }
     default:
-      throw new Error(`Unknown action: ${type}`)
+      return state
   }
 }