+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## Version 1.1.0 (2020/07/24)
+- Fixed filter placement in People and Sellers stories
+- Fixed filtering in Sellers story
+- Removed CSS hover on filters
+- Fixed size of selection column
+- Added scrollTop key to state to enable programatic control of body scrolling
+- Added VSCROLL action when the table body is scrolled vertically
+- Added icon to distinguish sortable and non-sortable columns
+- Fixed crash when attempting to sort in Basic table and Table with autosizing stories
+- upgraded dependencies
+- Improved log readability in samples
+
 ## Version 1.0.3 (2020/05/04)
 
 - Fixed incorrect use of forwardRef